Added standby mode script
